+++ .gitignore
@@ -8,10 +8,11 @@ node_modules
 
 # testing
 /coverage
+e2e_tests/cypress/videos/
 
 # production
 build
-e2e_tests/cypress/videos/
+data
 
 # misc
 .DS_Store
