+++ client/.eslintrc.js
@@ -1,6 +1,7 @@
 module.exports = {
   'env': {
     'browser': true,
+    'node': true,
     'es6': true,
     'jest/globals': true,
     'cypress/globals': true
