+++ client/src/index.js
@@ -5,7 +5,6 @@ import {
   BrowserRouter as Router
 } from 'react-router-dom'
 
-
 ReactDOM.render(
   <Router basename={process.env.PUBLIC_URL}>
     <App />
