+++ e2e_tests/package.json
@@ -6,7 +6,7 @@
     "test:e2e": "cypress run",
     "start:server": "cd ../server && npm start",
     "start:client": "cd ../client && npm start",
-    "start:prod": "cd ../client && npm run build && cp -r build/ ../server/build && cd ../server && npm start"
+    "start:prod": "cd ../client && npm run build && cp -r build/ ../server/build && cd ../server && npm run start:prod"
   },
   "devDependencies": {
     "cypress": "^6.5.0",
