+++ mongo/docker-compose.yml
@@ -1,0 +1,10 @@
+version: '3.7'
+
+services:
+  mongo:
+    image: mongo:bionic
+    volumes:
+      - ./mongo/data:/data/db
+    restart: unless-stopped
+    ports:
+      - 27017:27017
