+++ server/utils/config.js
@@ -2,11 +2,12 @@ if (process.env.NODE_ENV !== 'production') {
     require('dotenv').config()
   }
   
-  let PORT = process.env.PORT
-  let MONGODB_URI = process.env.MONGODB_URI
+  let PORT = process.env.PORT || 3001
+  let MONGODB_URI = 'mongodb://localhost:27017/dev_db'
+  //let MONGODB_URI = 'mongodb://kierratysavustin_db:27017/prod_db'
   
-  if (process.env.NODE_ENV === 'test') {
-    MONGODB_URI = process.env.TEST_MONGODB_URI
+  if (process.env.NODE_ENV === 'production') {
+    MONGODB_URI = 'mongodb://kierratysavustin_db:27017/prod_db'
   }
   
   module.exports = {
